add new file configuration for server and start to writing new api
